+++ packages/spacecat-shared-data-access/test/fixtures/index.fixtures.js
@@ -17,6 +17,8 @@ import configurations from './configurations.fixture.js';
 import experiments from './experiments.fixture.js';
 import importJobs from './import-jobs.fixture.js';
 import importUrls from './import-urls.fixture.js';
+import scrapeJobs from './scrape-jobs.fixture.js';
+import scrapeUrls from './scrape-urls.fixture.js';
 import keyEvents from './key-events.fixture.js';
 import opportunities from './opportunities.fixture.js';
 import organizations from './organizations.fixture.js';
@@ -46,4 +48,6 @@ export default {
   siteTopPages,
   sites,
   suggestions,
+  scrapeJobs,
+  scrapeUrls,
 };