+++ packages/spacecat-shared-data-access/test/unit/util.js
@@ -72,7 +72,7 @@ export const createElectroMocks = (Model, record) => {
         exclude: [
           'apiKey', 'audit', 'configuration', 'experiment',
           'importJob', 'importUrl', 'keyEvent', 'latestAudit', 'organization',
-          'opportunity', 'site', 'siteCandidate', 'siteTopPage', 'suggestion', 'asyncJob',
+          'opportunity', 'site', 'siteCandidate', 'siteTopPage', 'suggestion', 'asyncJob', 'scrapeJob', 'scrapeUrl',
         ],
       },
     },