+++ packages/spacecat-shared-rum-api-client/src/functions/form-vitals.js
@@ -10,12 +10,12 @@
  * governing permissions and limitations under the License.
  */
 
-import { DataChunks } from '@adobe/rum-distiller';
+import { DataChunks, facets } from '@adobe/rum-distiller';
 import trafficAcquisition from './traffic-acquisition.js';
-import { generateKey, DELIMITER, loadBundles } from '../utils.js';
+import { DELIMITER, generateKey, loadBundles } from '../utils.js';
 
 const METRICS = ['formview', 'formengagement', 'formsubmit'];
-const CHECKPOINTS = ['viewblock', 'click', 'fill', 'formsubmit', 'navigate', 'viewmedia'];
+const CHECKPOINTS = ['viewblock', 'click', 'fill', 'formsubmit', 'navigate', 'viewmedia', 'experiment'];
 const KEYWORDS_TO_FILTER = ['search'];
 
 function initializeResult(url) {
@@ -37,7 +37,7 @@ function filterEvents(bundles) {
         return false;
       }
 
-      if (event.checkpoint === 'navigate') {
+      if (event.checkpoint === 'navigate' || event.checkpoint === 'experiment') {
         return true;
       }
 
@@ -79,6 +79,18 @@ const metricFns = {
 
 function findByUrl(formVitals, url) {
   return Object.values(formVitals).find((item) => item.url === url);
+}
+
+function getAllURLWithExperiment(bundles) {
+  const dataChunks = new DataChunks();
+  loadBundles(bundles, dataChunks);
+  const { checkpoint } = facets;
+  dataChunks.addFacet('checkpoint', checkpoint);
+  return new Set(
+    dataChunks.facets.checkpoint
+      .filter((cp) => cp.value === 'experiment')
+      .flatMap((_checkpoint) => _checkpoint.entries.map((_entry) => _entry.url)),
+  );
 }
 
 function populateFormsInternalNavigation(bundles, formVitals) {
@@ -137,6 +149,10 @@ function findFormCTAForInternalNavigation(bundles, formVitals) {
 
 function containsFormVitals(row) {
   return METRICS.some((metric) => Object.keys(row[metric]).length > 0);
+}
+
+function isUnderExperiment(row, experimentUrls) {
+  return experimentUrls.has(row.url);
 }
 
 function getParentPageVitalsGroupedByIFrame(bundles, dataChunks, iframeParentMap) {
@@ -275,10 +291,14 @@ function handler(bundles) {
     iframeParentMapWithoutDuplicates,
   );
 
+  const experimentUrls = getAllURLWithExperiment(bundlesWithFilteredEvents);
+
   // populate internal navigation data
   populateFormsInternalNavigation(bundles, formVitals);
   // filter out pages with no form vitals
-  const filteredFormVitals = Object.values(formVitals).filter(containsFormVitals);
+  const filteredFormVitals = Object.values(formVitals).filter(
+    (formVital) => containsFormVitals(formVital) && !isUnderExperiment(formVital, experimentUrls),
+  );
   findFormCTAForInternalNavigation(bundles, filteredFormVitals);
 
   const updatedFormVitals = filteredFormVitals.map((formVital) => {