+++ packages/spacecat-shared-content-client/CHANGELOG.md
@@ -1,3 +1,45 @@
+# [@adobe/spacecat-shared-content-client-v1.7.9](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.7.8...@adobe/spacecat-shared-content-client-v1.7.9) (2025-06-19)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-helix-content-sdk to v1.4.15 ([#808](https://github.com/adobe/spacecat-shared/issues/808)) ([ae211b9](https://github.com/adobe/spacecat-shared/commit/ae211b9db3692ea53b884de10682ef203ba45877))
+
+# [@adobe/spacecat-shared-content-client-v1.7.8](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.7.7...@adobe/spacecat-shared-content-client-v1.7.8) (2025-06-14)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#802](https://github.com/adobe/spacecat-shared/issues/802)) ([fc2cb47](https://github.com/adobe/spacecat-shared/commit/fc2cb47183948833f5b0a411ae78d1649e747a17))
+
+# [@adobe/spacecat-shared-content-client-v1.7.7](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.7.6...@adobe/spacecat-shared-content-client-v1.7.7) (2025-06-07)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#794](https://github.com/adobe/spacecat-shared/issues/794)) ([756cc67](https://github.com/adobe/spacecat-shared/commit/756cc67513540b0fc69ead1b90345891fef64793))
+
+# [@adobe/spacecat-shared-content-client-v1.7.6](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.7.5...@adobe/spacecat-shared-content-client-v1.7.6) (2025-06-07)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-helix-content-sdk to v1.4.13 ([#793](https://github.com/adobe/spacecat-shared/issues/793)) ([0bb1881](https://github.com/adobe/spacecat-shared/commit/0bb18813a6e82dfadc5babe0ba5ddcabf328885b))
+
+# [@adobe/spacecat-shared-content-client-v1.7.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.7.4...@adobe/spacecat-shared-content-client-v1.7.5) (2025-06-06)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-helix-content-sdk to v1.4.12 ([#791](https://github.com/adobe/spacecat-shared/issues/791)) ([89249ef](https://github.com/adobe/spacecat-shared/commit/89249ef164c1c1fcc5b9d2a22d58b40d93785822))
+
+# [@adobe/spacecat-shared-content-client-v1.7.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.7.3...@adobe/spacecat-shared-content-client-v1.7.4) (2025-05-31)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#779](https://github.com/adobe/spacecat-shared/issues/779)) ([07f8cce](https://github.com/adobe/spacecat-shared/commit/07f8cce73e33bfb9c61fe14f2ef28012b872437d))
+
 # [@adobe/spacecat-shared-content-client-v1.7.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.7.2...@adobe/spacecat-shared-content-client-v1.7.3) (2025-05-26)