+++ packages/spacecat-shared-utils/CHANGELOG.md
@@ -1,3 +1,38 @@
+# [@adobe/spacecat-shared-utils-v1.40.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.39.1...@adobe/spacecat-shared-utils-v1.40.0) (2025-06-18)
+
+
+### Features
+
+* expose getStaticContent function ([#809](https://github.com/adobe/spacecat-shared/issues/809)) ([33ad2ed](https://github.com/adobe/spacecat-shared/commit/33ad2ed2a01c9df4700932befd839c77c59cc504))
+
+# [@adobe/spacecat-shared-utils-v1.39.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.39.0...@adobe/spacecat-shared-utils-v1.39.1) (2025-06-14)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#802](https://github.com/adobe/spacecat-shared/issues/802)) ([fc2cb47](https://github.com/adobe/spacecat-shared/commit/fc2cb47183948833f5b0a411ae78d1649e747a17))
+
+# [@adobe/spacecat-shared-utils-v1.39.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.38.6...@adobe/spacecat-shared-utils-v1.39.0) (2025-06-12)
+
+
+### Features
+
+* introduce isValidHelixPreviewUrl in context of preflight audits ([#798](https://github.com/adobe/spacecat-shared/issues/798)) ([9e3fdd5](https://github.com/adobe/spacecat-shared/commit/9e3fdd5887e5b02e5a9248b4732f00f4bfa84f81))
+
+# [@adobe/spacecat-shared-utils-v1.38.6](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.38.5...@adobe/spacecat-shared-utils-v1.38.6) (2025-06-07)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#794](https://github.com/adobe/spacecat-shared/issues/794)) ([756cc67](https://github.com/adobe/spacecat-shared/commit/756cc67513540b0fc69ead1b90345891fef64793))
+
+# [@adobe/spacecat-shared-utils-v1.38.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.38.4...@adobe/spacecat-shared-utils-v1.38.5) (2025-05-31)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#779](https://github.com/adobe/spacecat-shared/issues/779)) ([07f8cce](https://github.com/adobe/spacecat-shared/commit/07f8cce73e33bfb9c61fe14f2ef28012b872437d))
+
 # [@adobe/spacecat-shared-utils-v1.38.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.38.3...@adobe/spacecat-shared-utils-v1.38.4) (2025-05-24)